Remove resync-period (#8)
The hourly resync causes a large CPU spike for pods because we reset the metric and
reevaluate the whole list of pods for every event (the resync resubmits every cached
pod as an update event). We could address this by implementing our own 'resync' that
runs updateMetrics once every hour and that would achieve the same thing less
expensively.

However, I'm sceptical that there's much value in resyncing at all. Pod events happen
pretty frequently anyway. Even in our quietest clusters it seems like a new pod starts
at least every 6 hours. Namespaces are much more infrequent but even so I'm not sure
exactly what we expect could go wrong with updateMetrics that we think might be solved
by calling it again some time later. And even if we do have a bug that unpredictably
misses metrics, maybe it's better to surface it rather than relying on a fail safe to
gloss over it on an hourly basis.